changelayoutaction wrapper
