+++ coopihc/space/utils.py
@@ -1,7 +1,6 @@
 from coopihc.space.Space import Space
 import numpy
-import gym
-import functools
+
 
 
 def remove_prefix(text, prefix):
@@ -83,231 +82,3 @@ def multidiscrete_space(iterable_possible_values, dtype=numpy.int16):
     return Space([numpy.array(i).astype(dtype) for i in iterable_possible_values])
 
 
-def partialclass(cls, *args):
-    class ChangeLayoutAction(cls):
-        __init__ = functools.partialmethod(
-            cls.__init__,
-            *args,
-        )
-
-    return ChangeLayoutAction
-
-
-######## ================= Below is untested code, and should replace the class below with same name.
-
-# class _ChangeLayoutAction(gym.ActionWrapper):
-#     def __init__(self, env, spaces, wrapperflag):
-#         change_action = []
-#         for sp, wflag in zip(spaces, wrapperflag):
-#             if not wflag:
-#                 change_action.append([None])
-#                 continue
-#             change_action.append([sp.range])
-
-#         self.change_action = change_action
-#         super().__init__(env)
-
-#     def action(self, action):
-#         _action = []
-#         for n, a in enumerate(action):
-#             if change_action[n] == [None]:
-#                 _action.append(a)
-#             else:
-#                 _action.append(change_action[n][a])
-#         return _action
-
-
-class _ChangeLayoutAction(gym.ActionWrapper):
-    """_ChangeLayoutAction
-
-    Class prototype
-
-    :param gym: [description]
-    :type gym: [type]
-    """
-
-    def __init__(self, spaces, wrapperflag, env):
-        super().__init__(env)
-
-    def action(self, action):
-        return action
-
-
-class _ChangeLayoutObservation(gym.ObservationWrapper):
-    """_ChangeLayoutObservation
-
-    Class prototype
-
-    :param gym: [description]
-    :type gym: [type]
-    """
-
-    def __init__(self, spaces, wrapperflag, env):
-        super().__init__(env)
-
-    def observation(self, observation):
-        return observation
-
-
-class RLConvertor:
-    """Help convert Bundle to RL Envs
-
-    Helper class to convert Bundles to Reinforcement Learning environments
-
-    :param interface: name of the API to which bundles is converted to, defaults to "gym"
-    :type interface: str, optional
-    """
-
-    def __init__(self, interface="gym", **kwargs):
-        self.interface = interface
-        if self.interface != "gym":
-            raise NotImplementedError
-        else:
-            style = kwargs.get("style")
-            if style != "SB3":
-                raise NotImplementedError
-
-    def get_spaces_and_wrappers(self, spaces, mode):
-        """get_spaces_and_wrappers
-
-        Main method to call.
-
-        Return converted spaces, a potential wrapper and wrapper flags (True if space needs wrapper component). Currently, the only space that needs a wrapper potentially is DIscrete Space (when the start action is not 0 and actions not equally unit-spaced)
-
-        :param spaces: bundle spaces
-        :type spaces: list(Spaces`coopihc.spaces.Space.Space`)
-        :param mode: 'action' or 'observation'
-        :type mode: string
-        :return: spaces, wrapper, wrapperflag
-        :rtype: tuple([gym.spaces], [gym.env wrapper], [bool])
-        """
-        spaces, wrapperflag = self._convert(spaces)
-        if mode == "action":
-            wrapper = self._get_action_wrapper(spaces, wrapperflag)
-        elif mode == "observation":
-            wrapper = self._get_observation_wrapper(spaces, wrapperflag)
-        return spaces, wrapper, wrapperflag
-
-    def _get_info(self, spaces):
-        """_get_info
-
-        gather intel on bundle spaces
-
-        :param spaces: bundle spaces
-        :type spaces: list(Spaces`coopihc.spaces.Space.Space`)
-
-        """
-        # Check what mix spaces is
-
-        discrete = False
-        continuous = False
-        multidiscrete = 0
-        for space in spaces:
-            if space.continuous:
-                continuous = True
-            else:
-                discrete = True
-                multidiscrete += len(space)
-            if not isinstance(space, Space):
-                raise NotASpaceError(
-                    "All items in spaces should be of type Space, but space is of type {} ".format(
-                        type(space).__name__
-                    )
-                )
-
-        # Spaces -> collection of Discrete Space
-        if discrete:
-            # Check which spaces will need an actionwrapper
-            spacewrapper_flag = []
-            for space in spaces:
-                if multidiscrete > 1:
-                    for _space in space:
-                        if not (
-                            _space.range[0].squeeze().tolist()[0] == 0
-                            and numpy.mean(  # From gym 0.21, there is an option to offset the start, but only for discrete
-                                numpy.diff(_space.range)
-                            )
-                            == 1.0
-                            and numpy.std(numpy.diff(_space.range)) == 0.0
-                        ):
-                            spacewrapper_flag.append(True)
-                        else:
-                            spacewrapper_flag.append(False)
-                else:
-                    _space = space
-                    if not (
-                        _space.range[0].squeeze().tolist()[0] == 0
-                        and numpy.mean(  # From gym 0.21, there is an option to offset the start, but only for discrete
-                            numpy.diff(_space.range)
-                        )
-                        == 1.0
-                        and numpy.std(numpy.diff(_space.range)) == 0.0
-                    ):
-                        spacewrapper_flag.append(True)
-                    else:
-                        spacewrapper_flag.append(False)
-        else:
-            spacewrapper_flag = [False for space in spaces]
-
-        return discrete, continuous, multidiscrete, spacewrapper_flag
-
-
-class GymConvertor(RLConvertor):
-    """Help convert Bundle to RL Gym Envs
-
-    Helper class to convert Bundles to Gym Reinforcement Learning environments, with a certain library in mind. (Not all libraries may support all action and observation spaces.)
-
-    :param style: lib name
-    :type interface: str, optional
-    """
-
-    def __init__(self, style="SB3"):
-        super().__init__(interface="gym", style=style)
-
-    def _convert(self, spaces):
-        discrete, continuous, multidiscrete, spacewrapper_flag = self._get_info(spaces)
-        # Spaces -> mix of discrete and continuous Space
-        if continuous and discrete:
-            raise NotImplementedError(
-                "Mixing continuous and discrete Spaces requires a gym.spaces.Tuple of gym.spaces.Dict conversion, which is not implemeted yet. Consider doing it yourself an opening a PR on the project's github"
-            )
-        if discrete:
-            # Deal with one-unit collections
-            if multidiscrete == 1:
-                return gym.spaces.Discrete(spaces[0].N), spacewrapper_flag
-            else:
-                if len(spaces) == 1:
-                    return (
-                        gym.spaces.MultiDiscrete([s.N for s in spaces[0]]),
-                        spacewrapper_flag,
-                    )
-                else:
-                    return (
-                        gym.spaces.MultiDiscrete([s.N for s in spaces]),
-                        spacewrapper_flag,
-                    )
-        # Spaces -> collection of continuous Space
-        if continuous:
-            if len(spaces) != 1:
-                raise NotImplementedError(
-                    "Mixing continuous spaces requires a gym.spaces.Tuple of gym.spaces.Dict conversion, which is not implemeted yet. Consider doing it yourself an opening a PR on the project's github"
-                )
-            else:
-                return gym.spaces.Box(
-                    spaces[0].low,
-                    spaces[0].high,
-                    dtype=spaces[0].dtype,
-                    # seed=spaces[0].seed,   Starting from some gym version, this is possible
-                ), [False]
-
-    def _get_action_wrapper(self, spaces, wrapperflag):
-        if True not in wrapperflag:
-            return None
-        else:
-            return partialclass(_ChangeLayoutAction, spaces, wrapperflag)
-
-    def _get_observation_wrapper(self, spaces, wrapperflag):
-        if True not in wrapperflag:
-            return None
-        else:
-            return partialclass(_ChangeLayoutObservation, spaces, wrapperflag)
